+++ pkg/external/api/export_feeds_soak_test.go
@@ -1,14 +1,13 @@
 //go:build soak
 // +build soak
 
-package feed_test
+package api_test
 
 import (
 	"os"
 	"testing"
 
-	"github.com/SafetyCulture/safetyculture-exporter/pkg/external/api"
-	"github.com/SafetyCulture/safetyculture-exporter/pkg/internal/config"
+	"github.com/SafetyCulture/safetyculture-exporter/pkg/httpapi"
 	"github.com/SafetyCulture/safetyculture-exporter/pkg/internal/feed"
 
 	"github.com/stretchr/testify/assert"
@@ -19,10 +18,11 @@ func TestIntegrationDbSoakExportFeeds_should_successfully_export_with_significan
 	assert.NoError(t, err)
 	exporter.AutoMigrate = true
 
-	apiClient := api.NewClient(os.Getenv("TEST_API_HOST"), os.Getenv("TEST_ACCESS_TOKEN"))
+	apiClient := httpapi.NewClient(os.Getenv("TEST_API_HOST"), os.Getenv("TEST_ACCESS_TOKEN"))
 
-	cfg := &config.ExporterConfiguration{}
-	cfg.AccessToken = "token-123"
+	cfg := &feed.ExporterFeedCfg{
+		AccessToken: "token-123",
+	}
 
 	exporterApp := feed.NewExporterApp(apiClient, nil, cfg)
 	err = exporterApp.ExportFeeds(exporter)
