+++ pkg/external/api/utils_test.go
@@ -2,6 +2,7 @@ package api_test
 
 import (
 	"bytes"
+	"fmt"
 	"io"
 	"log"
 	"os"
@@ -13,6 +14,7 @@ import (
 	exporterAPI "github.com/SafetyCulture/safetyculture-exporter/pkg/external/api"
 	"github.com/SafetyCulture/safetyculture-exporter/pkg/httpapi"
 	"github.com/SafetyCulture/safetyculture-exporter/pkg/internal/feed"
+	"github.com/gofrs/uuid"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -118,3 +120,52 @@ func getTemporaryCSVExporterWithMaxRowsLimit(maxRowsPerFile int) (*feed.CSVExpor
 }
 
 var dateRegex = regexp.MustCompile(`(?m)(-?(?:[1-9][0-9]*)?[0-9]{4})-(1[0-2]|0[1-9])-(3[01]|0[1-9]|[12][0-9])T(2[0-3]|[01][0-9]):([0-5][0-9]):([0-5][0-9])(\.[0-9]+)?(\+|Z)(2[0-3]|[01][0-9])?:?([0-5][0-9])?`)
+
+// getTestingSQLExporter creates a temporary DB on the target SQL Database
+func getTestingSQLExporter() (*feed.SQLExporter, error) {
+	dialect := os.Getenv("TEST_DB_DIALECT")
+	connectionString := os.Getenv("TEST_DB_CONN_STRING")
+
+	exporter, err := feed.NewSQLExporter(dialect, connectionString, true, "")
+	if err != nil {
+		return nil, err
+	}
+
+	dbName := strings.ReplaceAll(fmt.Sprintf("iaud_exporter_%s", uuid.Must(uuid.NewV4()).String()), "-", "")
+
+	switch dialect {
+	case "postgres", "mysql", "sqlserver":
+		dbResp := exporter.DB.Exec(fmt.Sprintf(`CREATE DATABASE %s;`, dbName))
+		err = dbResp.Error
+	case "sqlite":
+		return exporter, nil
+	default:
+		return nil, fmt.Errorf("Invalid DB dialect %s", dialect)
+	}
+	if err != nil {
+		return nil, err
+	}
+
+	connectionString = strings.Replace(connectionString, "safetyculture_exporter_db", dbName, 1)
+	connectionString = strings.Replace(connectionString, "master", dbName, 1)
+
+	return feed.NewSQLExporter(dialect, connectionString, true, "")
+}
+
+// getTemporaryCSVExporterWithRealSQLExporter creates a CSV exporter that writes a temporary folder
+// but also uses a real DB as an intermediary
+func getTemporaryCSVExporterWithRealSQLExporter(sqlExporter *feed.SQLExporter) (*feed.CSVExporter, error) {
+	dir, err := os.MkdirTemp("", "export")
+	if err != nil {
+		return nil, err
+	}
+
+	exporter, err := feed.NewCSVExporter(dir, "", 100000)
+	if err != nil {
+		return nil, err
+	}
+
+	exporter.SQLExporter = sqlExporter
+
+	return exporter, err
+}
